# docs/panel-api.openapi.yaml
openapi: 3.0.3
info:
  title: SIGE Superadmin Panel API
  version: 1.0.0
  description: Internal API for the central superadmin panel.
servers:
  - url: http://localhost:3001
security:
  - PanelBearer: []
components:
  securitySchemes:
    PanelBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /api/superadmin/tenants:
    get:
      summary: List tenants
      responses:
        '200':
          description: Tenant list
    post:
      summary: Register tenant (URLs derived from slug SaaS hosts when omitted)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [slug, controlToken]
              properties:
                slug:
                  type: string
                displayName:
                  type: string
                controlToken:
                  type: string
                baseUrl:
                  type: string
                  description: Optional override; default https://erp.<slug>.findspo.com
                webBaseUrl:
                  type: string
                  description: Optional override; default https://www.<slug>.findspo.com
                provisionDomains:
                  type: boolean
                  description: If true, run full-auto DNS/Ingress/cert after create
      responses:
        '201':
          description: Created tenant
...